% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{all.algorithms}
\alias{all.algorithms}
\title{Library of Machine learning procedures}
\usage{
\method{all}{algorithms}()
}
\value{
a character vector with all prediction algorithms  supported by stackBagg
}
\description{
Library of Machine Learning procedures
}
