% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datageneration.R
\name{datagenPaper}
\alias{datagenPaper}
\title{Simulation settings from Gonzalez Ginestet et al. (2019+). "Ensemble IPCW Bagging bagging: a case study in the HIV care registry"}
\usage{
datagenPaper(J, n, frac.train, tao = 26.5, simulation, scenario)
}
\arguments{
\item{J}{number of simulated data sets}

\item{n}{number of sample size}

\item{frac.train}{percentange train data set. A number between 0 and 1.}

\item{tao}{time point of interest}

\item{simulation}{study indicator. It takes on 1 and 2.}

\item{scenario}{scenario indicator. It takes on 1, 2, 3 and 4.}
}
\value{
A list with the following elements: \describe{ \item{train.data}{ simulated train data set}
\item{test.data}{simulated  test data set} } and the true AUC
}
\description{
Function that generates data of the different simulation studies
presented in the accompanying paper. This function requires the
  \code{gam} package to be installed.
}
\examples{
DT=datagenPaper(J = 1 , n = 1250, frac.train = 0.8, simulation=1, scenario=4)
}
