% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ens.all.algorithms}
\alias{ens.all.algorithms}
\title{Library of Machine learning procedures}
\usage{
ens.all.algorithms()
}
\value{
a character vector with all prediction algorithms  supported by ensBagg
}
\description{
Library of Machine Learning procedures
}
