% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{EnsBagg-internal}
\alias{EnsBagg-internal}
\alias{optimun_auc_coef}
\alias{risk_auc}
\alias{MLprocedures}
\alias{ML_list}
\alias{MLprocedures_natively}
\alias{ML_list_natively}
\alias{grid_parametersDataHIV}
\title{Internal EnsBagg functions}
\format{An object of class \code{list} of length 8.}
\usage{
optimun_auc_coef(coef_init, lambda, data, Z, tao)

risk_auc(par, lambda, Z, data, tao)

MLprocedures(traindata, testdata, fmla, xnam, xnam.factor, xnam.cont,
  xnam.cont.gam, tuneparams, i)

ML_list

MLprocedures_natively(traindata, testdata, fmla, xnam, xnam.factor,
  xnam.cont, xnam.cont.gam, tuneparams)

ML_list_natively

grid_parametersDataHIV(fmla, xnam, data)
}
\arguments{
\item{coef_init}{starting values for the coefficients}

\item{lambda}{penalization term. It is a positive scalar.}

\item{data}{A data frame that contains at least: ttilde, delta, wts}

\item{Z}{a matrix that contains the predictions. Each column represents a single marker.}

\item{tao}{time point of interest}

\item{par}{a vector of coefficients/weights. Its length must be equal to the number of predictions included in Z}

\item{traindata}{training data set}

\item{testdata}{validation/test data set}

\item{fmla}{formula object ex. "E ~ x1+x2"}

\item{xnam}{a vector with the covariates names considered in the modeling}

\item{tuneparams}{a list of tune parameters for each machine learning procedure}

\item{i}{sample selected by bootstrap}

\item{lambda}{penalization term. It is a positive scalar.}

\item{Z}{a matrix that contains the predictions. Each column represents a single marker.}

\item{data}{A data frame  that constains at least: ttilde= time to event, delta=event type, wts= IPC weights}

\item{tao}{time point of interest}

\item{traindata}{training data set}

\item{testdata}{validation/test data set}

\item{fmla}{formula object ex. "E ~ x1+x2"}

\item{tuneparams}{a list of tune parameters for each machine learning procedure}

\item{fmla}{formula object ex. "E ~ x1+x2"}

\item{data}{a training data set}
}
\value{
a vector with the optimal AUC value and the optimal coefficient

1-AUC

a matrix of predictions where each column is the prediction of each algorithm based on the testdata

a list of Machine Learning functions

a matrix of predictions where each column is the prediction of each algorithm based on the testdata

a list of Machine Learning functions

a list with a grid of values for each hyperparameter
}
\description{
Internal EnsBagg helper functions

Compute the risk of missclassifying an individual using as a marker a single prediction or weighted linear combination of several predictions (1-AUC)

Predictions based on a library of Machine Learning procedures

Library of Machine Learning procedures

Predictions based on those Machine Learning procedures in the library that allow for weights to be specified as an argument of the R function. No bagging occurs. This group of algorithms is denoted as  Native Weights

Library of Machine Learning procedures that allows for weights

A grid of values for hyperparameters used in the Real Data Application: InfCareHIV Register. This grid of values isan argument in the tuning parameter function tune_parameter_ml.R
}
\details{
These functions are not intended for use by users.
}
\keyword{datasets}
