% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsBagg-package.R
\docType{package}
\name{EnsBagg-package}
\alias{EnsBagg}
\alias{EnsBagg-package}
\title{EnsBagg: Ensemble IPCW Bagging}
\description{
Estimation the risk of an event at a specific time in the presence of censored data with competing risk using a collection of 
             machine learning algorithms.
             Each algorithm is trained on a IPC-weighted bootstrap sample from the training set and predictions are obtained for the validation set.
             The Ensemble IPCW Bagging prediction is given by the weighted linear combination of the prediction of each algorithm.
             The area under the IPCW time-varying receiver operator curve (AUC) is used for evaluating
             the predictive performance and calibrating the ensemble and each single algorithm.
}
\author{
\strong{Maintainer}: Pablo Gonzalez Ginestet \email{pablo.gonzalez.ginestet@ki.se} (https://staff.ki.se/people/pabgon)

Authors:
\itemize{
  \item Erin Gabriel \email{erin.gabriel@ki.se} (https://staff.ki.se/people/erigab)
  \item Julian Wolfson \email{julianw@umn.edu} (https://sites.google.com/umn.edu/julianw)
}

}
\keyword{internal}
