% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{ipcw_auc}
\alias{ipcw_auc}
\title{IPC Weighted AUC Loss Function}
\usage{
ipcw_auc(T, delta, marker, cause, wts, tao)
}
\arguments{
\item{T}{vector of (censored) event-times}

\item{delta}{vector of event indicators at the corresponding value of the vector T. Censored observations must be denoted by the value 0.}

\item{marker}{vector of the marker values for which we want to compute the time-dependent ROC curve. the function assumes that larger values of the marker are associated with higher risks of events}

\item{cause}{value of the event indicator (the non-censored observation) that represents the event of interest for which we aim to compute the time-dependent ROC curve.}

\item{wts}{IPC weights}

\item{tao}{evaluation time point of interest}
}
\value{
value of the AUC at tao
}
\description{
Compute time-varying IPCW AUC to account for censoring and competing risks.
}
