% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipcw_bagging.R
\name{ipcw_ensbagg}
\alias{ipcw_ensbagg}
\title{Algorithm 2: Ensemble IPCW Bagging}
\usage{
ipcw_ensbagg(folds, MLprocedures, fmla, tuneparams, tao, B = NULL, A,
  data, xnam, xnam.factor, xnam.cont, xnam.cont.gam, ens.library)
}
\arguments{
\item{folds}{Number of folds}

\item{MLprocedures}{\link{MLprocedures}}

\item{fmla}{formula object ex. "E ~ x1+x2"}

\item{tuneparams}{a list of tune parameters for each machine learning procedure}

\item{tao}{time point of interest}

\item{B}{number of bootstrap samples}

\item{data}{a training data set}

\item{xnam}{all covariates in the model}

\item{xnam.factor}{categorical variables include in the model}

\item{xnam.cont}{continous variables include in the model}

\item{xnam.cont.gam}{continous variables to be included in the smoothing operator gam::s(,df)}

\item{ens.library}{algorithms in the library}
}
\value{
a list with the predictions of each machine learning algorithm (id, predictions), the average AUC across folds for each of them, the optimal coefficients, an indicator if the optimization procedure has converged and the value of penalization term chosen
}
\description{
Obtain predictions
}
