% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipcw_bagging.R
\name{ipcw_genbagg}
\alias{ipcw_genbagg}
\title{Algorithm 1: General IPCW Bagging}
\usage{
ipcw_genbagg(fmla, tuneparams, MLprocedures, traindata, testdata, B, A,
  xnam, xnam.factor, xnam.cont, xnam.cont.gam, ens.library)
}
\arguments{
\item{fmla}{formula object ex. "E ~ x1+x2"}

\item{tuneparams}{a list of tune parameters for each machine learning procedure}

\item{MLprocedures}{\link{MLprocedures}}

\item{traindata}{a training data set}

\item{testdata}{a test data set}

\item{B}{number of bootstrap samples}

\item{xnam}{all covariates in the model}

\item{xnam.factor}{categorical variables include in the model}

\item{xnam.cont}{continous variables include in the model}

\item{xnam.cont.gam}{continous variables to be included in the smoothing operator gam::s(,df=)}

\item{ens.library}{algorithms in the library}
}
\value{
a matrix with the predictions on the test data set of each machine learning algorithm considered in \link{MLprocedures}
}
\description{
Obtain predictions
}
