% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametersSimulation.R
\name{parametersSimulation}
\alias{parametersSimulation}
\title{Parameter Values for the Simulation}
\usage{
parametersSimulation(folds, xnam, data, tao)
}
\arguments{
\item{folds}{number of folds}

\item{xnam}{a vector with the covariates names considered in the modeling}

\item{data}{a training data set}

\item{fmla}{formula object ex. "E ~ x1+x2"}
}
\value{
a list of values for each hyperparameter in each algorithm used in the simulation setup
}
\description{
Values for the hyperparameters used in the simulation setup
}
