% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_auc.R
\name{plot_ipcw_roc}
\alias{plot_ipcw_roc}
\title{Plot IPCW ROC curve}
\usage{
plot_ipcw_roc(time, delta, marker, wts, tao)
}
\arguments{
\item{time}{}

\item{tao}{evaluation time point of interest}
}
\value{
a plot IPCW ROC curve
}
\description{
Plot IPCW ROC curve for a prediction
}
