% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_auc.R
\name{plot_roc}
\alias{plot_roc}
\title{Plot IPCW ROC curve}
\usage{
plot_roc(time, delta, marker, wts = NULL, tao, method = "ipcw")
}
\arguments{
\item{time}{}

\item{tao}{evaluation time point of interest}

\item{method}{IPCW or discard}
}
\value{
a plot IPCW ROC curve
}
\description{
Plot IPCW ROC curve for a prediction
}
