% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackBagg-package.R
\docType{package}
\name{stackBagg-package}
\alias{stackBagg}
\alias{stackBagg-package}
\title{stackBagg: Stacked IPCW Bagging}
\description{
Estimation the risk of an event at a specific time in the presence of censored data with competing risk using a collection of 
             machine learning algorithms.
             Each algorithm is trained on a IPC-weighted bootstrap sample from the training set and predictions are obtained for the validation set.
             The stacked IPCW Bagging prediction is given by the weighted linear combination of the prediction of each algorithm.
             The area under the IPCW time-dependent receiver operator curve (AUC) is used for evaluating
             the predictive performance and calibrating the stack and each single algorithm.
}
\author{
\strong{Maintainer}: Pablo Gonzalez Ginestet \email{pablo.gonzalez.ginestet@ki.se} (https://staff.ki.se/people/pabgon)

}
\keyword{internal}
