% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametersSimulation.R
\name{tune_lasso}
\alias{tune_lasso}
\title{Obtain the cross-validated lambda hyparameter for the LASSO}
\usage{
tune_lasso(folds, fmla, tao, data, xnam)
}
\arguments{
\item{folds}{number of folds}

\item{fmla}{formula object ex. "E ~ x1+x2"}

\item{tao}{time point of interest}

\item{data}{a training data set}
}
\value{
lambda to be used in the glmnet function
}
\description{
Obtain the lambda hyparameter for the LASSO using cross-validation
}
