% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_parameters_ml.R
\name{tune_params_ml}
\alias{tune_params_ml}
\title{Tuning parameter selection}
\usage{
tune_params_ml(gam_param, lasso_param, randomforest_param, knn_param,
  svm_param, nn_param, bart_param, folds, xnam, tao, data)
}
\arguments{
\item{gam_param}{a vector containing degree of freedom 3 and 4}

\item{lasso_param}{a grid of values for the shrinkage term}

\item{randomforest_param}{a two column matrix: first column denotes the num_trees parameter and the second column denotes the mtry parameter.}

\item{knn_param}{a grid of  positive integers values}

\item{svm_param}{a three column matrix:  first column denotes the cost parameter, second column the gamma and third column the kernel. kernel=1 denotes "radial" and kernel=2 denotes "linear".}

\item{nn_param}{a grid of positive integers values for the neurons}

\item{bart_param}{a three column matrix:  first column denotes the num_tree parameter, second column the k parameter and third column the q parameter.}

\item{folds}{number of folds}

\item{xnam}{vector with the names of the covariates to be included in the model}

\item{tao}{evaluation time point of interest}

\item{data}{data set that contains at least id, E , ttilde, delta, wts and covariates}
}
\value{
a list with the tune parameters selected using the IPCW AUC loss function.
}
\description{
Tuning parameter selection
}
