% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{ipcw_crossentropy}
\alias{ipcw_crossentropy}
\title{IPC Weighted Cross Entropy Loss Function}
\usage{
ipcw_crossentropy(par, Z, y, wts)
}
\arguments{
\item{par}{a vector of weights. Its length must be equal to the number of predictions included in Z}

\item{Z}{a matrix that contains the predictions. Each column represents a single marker.}

\item{y}{vector of response variable (binary).}

\item{wts}{IPC weights}
}
\description{
Compute weighted Cross Entropy Loss function for a single marker or a linear weighted combination of markers
}
