% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensBagg.R
\name{ensBagg}
\alias{ensBagg}
\title{Ensemble IPCW Bagging}
\usage{
ensBagg(train.data, test.data, xnam, tao, weighting, folds, ens.library,
  tuneparams = NULL, B = NULL)
}
\arguments{
\item{train.data}{a data.frame with at least the following variables: 
event-times (censored) in the first column,
event indicator in the second column and 
covariates/features that the user potentially want to use in building the preodiction model.
Censored observations must be denoted by the value 0. Main event of interest is denoted by 1.}

\item{test.data}{a data.frame with the same variables and names that the train.data}

\item{xnam}{vector with the names of the covariates to be included in the model}

\item{tao}{evaluation time point of interest}

\item{weighting}{Procedure to compute the inverse probability of censoring weights. Weighting="CoxPH" and weighting="CoxBoost" model the censoring by the Cox model and CoxBoost model respectively.}

\item{folds}{Number of folds}

\item{ens.library}{character vector of prediction algorithms. The prediction algorithms supported are: "ens.glm","ens.gam","ens.lasso","ens.randomForest","ens.svm","ens.bartMachine","ens.knn","ens.nn")}

\item{tuneparams}{a list of tune parameters for each machine learning procedure. Name them as gam_param, lasso_param, randomforest_param, svm_param, bart_param, knn_param, nn_param.
Default values are the same used for the simulation.}

\item{B}{number of bootstrap samples}
}
\value{
a list with the predictions of each machine learning algorithm, the average AUC across folds for each of them, the optimal coefficients,the weights ,an indicator if the optimization procedure has converged and the value of penalization term chosen
}
\description{
Main Algorithm
}
